start Model-containing framework for RunDirectory
RunDirectory is supposed to be a wrapper for everything run-related. The
rest of the filesystem should just as well not need to *exist* when
Engine kicks off a Job.

This means that RunDirectory should, ideally, grab the responsibility of
ensuring the model is copied correctly and that none is trying to access
(especially write or unlink!) anything outside itself, etc.

Simply put, all run filesystem IO is to go through RunDirectory.