add SourceResource -- a resource manager API for source code
(Not yet replaced old interface. Tomorrow?)

Might feel unnecessary, but it was starting to get messy with entangling
all on Model class. What is it doing there, anyway?

Idea is that this is a "pure" manager of source code/source-code like
I/O, which can deal with caching which the less pure Model-API's
(generally) shouldn't. Seriously, I really want to guarantee source
code input immutability and not just "hope" that file-like object hasn't
mutated from init read.

To not speak of whatever happens when different versions are accessed by
different APIs due to temporal separation... Fun.