refactor Job system to inherit threading.Thread
Yes, it always threads now. Hopefully easier to handle.

Also found the bug with freezes. Was a joining race condition + lack of
"correct" exit management.

When the main thread is watching a child process for its return code, it
is NOT A WISE IDEA to block until the thread joins! The thread can ONLY
join once the child has exited, which it can ONLY do once the main
thread has been called (i.e. the asyncio child watcher) which can ONLY
happen when it's not, you know, blocking...

Python has GIL. Without "import multiprocessing" a block is global, even
to the point of all subprocess signal propagation. EVERYTHING stalls.