Add workaround for problems with jupyter under rstudio in windows
In some cases the reporting crashes when running from rstudio in
Windows. A WinError 6 is triggered for stdout. This fix spoofs the name
of the Python executable so that jupyter_client believes it runs
pythonw which triggers jupyter_client to turn of stdout and stderr which
solves the problem. There might be cleaner ways of doing this, but this
would probably require patching of the dependency.