Speed up large dataset parsing by optimizing filtering and conversion
We perform filtering and conversion in at most two passes.

Conversion passes are interleaved with filtering passes to avoid converting
entries that get excluded anyway.

Conversion is slightly optimized by performing multiple-column
conversions with a single `apply` call by vectorizing `_convert_data_item`.

We also allow the user to control which columns should be converted in the
result dataframe when `raw=False`. We ensure that columns necessary for
numeric filtering are converted before that filtering pass, and
"unconvert" those afterwards if the user did not request for them to be
converted in result.