Speed up TSV-like dataset parsing for NONMEM
This prevents us from using this implementation for output parsing since the
separator used there is `sep=r'\s+'` which would merge consecutive tabs into a
single separator. However, we do not currently use this implementation to parse
output as it is slower than a direct call to `pd.read_table` (obviously).